Make EAN more tolerant
If you add spaces, it will ignore them; if you omit the number system it will assume 0, and if you omit the check digit it will calculate it for you. The printed characters are left alone (so if you have spaces in the input, the space will show up in the printed characters).